BFT: move shared peer & orderer block delivery client code to common
