+++ common/deliverclient/blocksprovider/bft_censorship_monitor.go
@@ -16,7 +16,7 @@ import (
 	"github.com/hyperledger/fabric-lib-go/common/flogging"
 	"github.com/hyperledger/fabric-protos-go/common"
 	"github.com/hyperledger/fabric-protos-go/orderer"
-	"github.com/hyperledger/fabric/internal/pkg/peer/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/orderers"
 	"github.com/pkg/errors"
 	"go.uber.org/zap/zapcore"
 )
