+++ common/deliverclient/blocksprovider/bft_censorship_monitor_factory.go
@@ -6,7 +6,7 @@ SPDX-License-Identifier: Apache-2.0
 
 package blocksprovider
 
-import "github.com/hyperledger/fabric/internal/pkg/peer/orderers"
+import "github.com/hyperledger/fabric/common/deliverclient/orderers"
 
 // BFTCensorshipMonitorFactory creates an instance of a BFTCensorshipMonitor. It is an implementation of the
 // CensorshipDetectorFactory interface which abstracts the creation of a BFTCensorshipMonitor.
