+++ common/deliverclient/blocksprovider/bft_censorship_monitor_factory_test.go
@@ -9,7 +9,7 @@ package blocksprovider_test
 import (
 	"testing"
 
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider"
 	"github.com/stretchr/testify/require"
 )
