+++ common/deliverclient/blocksprovider/bft_censorship_monitor_test.go
@@ -17,9 +17,9 @@ import (
 	"github.com/hyperledger/fabric-lib-go/common/flogging"
 	"github.com/hyperledger/fabric-protos-go/common"
 	"github.com/hyperledger/fabric-protos-go/orderer"
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider"
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider/fake"
-	"github.com/hyperledger/fabric/internal/pkg/peer/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider/fake"
+	"github.com/hyperledger/fabric/common/deliverclient/orderers"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 )
