+++ common/deliverclient/blocksprovider/fake/deliver_client_requester.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/hyperledger/fabric-protos-go/common"
 	"github.com/hyperledger/fabric-protos-go/orderer"
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider"
-	"github.com/hyperledger/fabric/internal/pkg/peer/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider"
 )
 
 type DeliverClientRequester struct {
