+++ common/deliverclient/blocksprovider/fake/dialer.go
@@ -4,7 +4,7 @@ package fake
 import (
 	"sync"
 
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider"
 	"google.golang.org/grpc"
 )
