+++ common/deliverclient/blocksprovider/fake/duration_exceeded_handler.go
@@ -4,7 +4,7 @@ package fake
 import (
 	"sync"
 
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider"
 )
 
 type DurationExceededHandler struct {
