+++ common/deliverclient/blocksprovider/fake/orderer_connection_source.go
@@ -4,8 +4,8 @@ package fake
 import (
 	"sync"
 
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider"
-	"github.com/hyperledger/fabric/internal/pkg/peer/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider"
 )
 
 type OrdererConnectionSource struct {
