+++ common/deliverclient/blocksprovider/fake/orderer_connection_source_factory.go
@@ -5,8 +5,8 @@ import (
 	"sync"
 
 	"github.com/hyperledger/fabric-lib-go/common/flogging"
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider"
-	"github.com/hyperledger/fabric/internal/pkg/peer/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider"
 )
 
 type OrdererConnectionSourceFactory struct {
