+++ common/deliverclient/blocksprovider/fake/updatable_block_verifier.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/hyperledger/fabric-protos-go/common"
 	"github.com/hyperledger/fabric/common/deliverclient"
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider"
 )
 
 type UpdatableBlockVerifier struct {
