+++ common/deliverclient/orderers/connection_factory_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 
 	"github.com/hyperledger/fabric-lib-go/common/flogging"
-	"github.com/hyperledger/fabric/internal/pkg/peer/orderers"
+	"github.com/hyperledger/fabric/common/deliverclient/orderers"
 	"github.com/stretchr/testify/require"
 )
