+++ core/deliverservice/gossip_block_handler_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/hyperledger/fabric-lib-go/common/flogging"
 	"github.com/hyperledger/fabric-protos-go/common"
-	"github.com/hyperledger/fabric/internal/pkg/peer/blocksprovider/fake"
+	"github.com/hyperledger/fabric/common/deliverclient/blocksprovider/fake"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 )
